refactor(frontend): split TaskProgressBar into StatusIcon and TaskRow sub-components
Extract TaskRow and StatusIcon into their own files under components/ (one
component per file), and move shared animation constants (animations.ts) and
the display-status helper (toDisplayStatus/DisplayStatus) alongside them.
