+++ autogpt_platform/frontend/src/app/(platform)/copilot/components/TaskProgressBar/animations.ts
@@ -1,0 +1,10 @@
+// ease-out-quad — Emil's default for entering elements.
+export const EASE_OUT = [0.25, 0.46, 0.45, 0.94] as const;
+
+// A newly in-progress loader waits for the completing task's tick + strikethrough
+// to play first, so each step reads as a sequence.
+export const LOADER_DELAY = 0.3;
+
+// Blur-bridge for the status icon swap (Emil's copy-button pattern).
+export const ICON_EXIT = 0.12;
+export const ICON_ENTER = 0.18;
